Attempt to simulate a reinit of socket connection
This should allow us to test behaviour following a restart of
the socket connection without a corresponding reset of the M4
core. However, the implementation of `Application_Connect`
apparently doesn't release all resources when you close the file
descriptor, so further work needs to be done to work out the best
way of simulating this behaviour.